requirements.txt
django
Pillow
python-dotenv
boto3 
django-storages